Add files via upload
1.	Objective:
The main goal is to apply image segmentation techniques, such as thresholding and edge detection, to extract features from objects in an image. The process focuses on dividing the image into parts to identify objects like apples and oranges and extract their relevant features. The goal is to identify, segment, and analyze objects for further image analysis or classification.
2.	Problem Statement:
Problem:
The problem is to segment an image containing multiple objects and extract useful features for object recognition. Techniques like thresholding and edge detection are used to separate objects from the background and extract features such as area, perimeter, and bounding 
Output:
The expected output is a transformed image that addresses the problem. For example, if the problem is edge detection, the output should be an image with highlighted edges. If the problem is thresholding, the final image should be binary, with distinct black and white regions.